# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ServiceRequestInfo:
    properties:
      approval_required:
        type: boolean
      assigned_to:
        type: integer
      assigned_to_name:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      department_name:
        type: string
      fulfilled_at:
        type: string
      id:
        type: integer
      pending_approvals:
        type: integer
      priority:
        type: string
      related_asset_id:
        type: integer
      related_asset_name:
        type: string
      requested_by:
        type: integer
      requested_by_name:
        type: string
      service_catalog_code:
        type: string
      service_catalog_id:
        type: integer
      service_catalog_name:
        type: string
      sr_number:
        type: string
      status:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Backend API implementing ISO/IEC 19770-10 Grade A++ RBAC model
  title: IT Asset & Compliance API
  version: "1.1"
paths:
  /approvals:
    get:
      parameters:
      - description: service_request atau change_request
        in: query
        name: entity_type
        required: true
        type: string
      - description: ID entitas
        in: query
        name: entity_id
        required: true
        type: integer
      responses: {}
      summary: List approval history untuk suatu entitas
      tags:
      - Approval
    post:
      responses: {}
      summary: Tambah approver ke workflow (service_request / change_request)
      tags:
      - Approval
...
  /service-catalog/{id}:
    delete:
      parameters:
      - description: Catalog Item ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Soft-delete item service catalog
      tags:
      - Service Catalog
    put:
      parameters:
      - description: Catalog Item ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Update item service catalog
      tags:
      - Service Catalog
  /service-requests:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Hanya request milik sendiri
        in: query
        name: my
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ServiceRequestInfo'
            type: array
      summary: List semua service requests
      tags:
      - Service Request
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      summary: Buat service request baru
      tags:
      - Service Request
  /service-requests/{id}:
    delete:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Soft-delete service request
      tags:
      - Service Request
    get:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Detail service request beserta approval history
      tags:
      - Service Request
    put:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Update service request (assignee, notes, dll)
      tags:
      - Service Request
  /service-requests/{id}/cancel:
    post:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Batalkan service request
      tags:
      - Service Request
  /service-requests/{id}/fulfill:
    post:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Tandai service request selesai dipenuhi
      tags:
      - Service Request
  /service-requests/{id}/start:
    post:
      parameters:
      - description: Service Request ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Mulai proses pemenuhan SR → in_fulfillment
      tags:
      - Service Request
...
swagger: "2.0"